Port TestChronoLocalDateTime to scalatest
